refactor: RacingGame과 CarDTO 의존성 분리
