+++ src/main/java/racingcar/dto/CarDTO.java
@@ -9,9 +9,13 @@ public class CarDTO {
     private final Name name;
     private final Position position;
 
-    public CarDTO(final Car car) {
-        this.name = car.getName();
-        this.position = car.getPosition();
+    private CarDTO(final Name name, final Position position) {
+        this.name = name;
+        this.position = position;
+    }
+
+    public static CarDTO from(final Car car) {
+        return new CarDTO(car.getName(), car.getPosition());
     }
 
     public String getName() {
